chore: Update deployment configuration for Heroku to use new environment variable names
Refactor the deployment configuration in the .github/workflows/deployment.yml file to use the new environment variable names for the server. This includes updating the variable names for MongoDB, secret, JWT key, Stripe secret key, email, password, and reset password key. The new variable names have been prefixed with "HD_" to indicate the changes. This update ensures that the server can access the correct environment variables when deployed to Heroku.